Refactor: 쪽지시험 컴포넌트 중 ExamOption 사용 부분 수정
- 쪽지시험 컴포넌트 중 ExamOption 사용 부분 수정
- 쪽지시험 컴포넌트 중 ExamOption 부분 구조 분해로 사용 가능하게 수정

Refactor: #43
